fix: student number was Integer. This caused problems in the register because the student number could be higher then the capacity of the Integer, so it was altered to Long.

+++ src/main/java/pt/isec/client/core/ClientManager.java
@@ -93,7 +93,7 @@ public class ClientManager implements IClientControllerContext, IClientThreadCon
     private String  userEmail;
     private String  userName;
     private Integer userId;
-    private Integer studentNumber;
+    private Long studentNumber;
 
     // Directory config
     private final int    directoryUdpPort;
@@ -291,7 +291,7 @@ public class ClientManager implements IClientControllerContext, IClientThreadCon
     }
 
     @Override
-    public Integer getStudentNumber() {
+    public Long getStudentNumber() {
         return studentNumber;
     }
 
@@ -323,8 +323,8 @@ public class ClientManager implements IClientControllerContext, IClientThreadCon
     }
 
     @Override
-    public void setStudentNumber(Integer number) {
-        Integer old = this.studentNumber;
+    public void setStudentNumber(Long number) {
+        Long old = this.studentNumber;
         this.studentNumber = number;
         pcs.firePropertyChange(PROP_STUDENT_NUMBER, old, number);
     }
